refactor: orgnize components
