+++ src/components/pages/details/anime-progress.tsx
