+++ src/components/pages/details/character.tsx
