+++ src/components/pages/details/list.tsx
