+++ src/components/pages/details/progress.tsx
