+++ src/components/pages/details/relations.tsx
