+++ src/components/pages/feed/entry.tsx
