+++ src/components/pages/home/counter.tsx
