+++ src/components/pages/home/hero.tsx
