+++ src/components/pages/user/cast-favorite-card.tsx
