+++ src/components/pages/user/favorite-card.tsx
