+++ src/components/shared/cards/core-team.tsx
