+++ src/components/shared/cards/details.tsx
